fix(ExpAdjuster): Do not set values that are unchanged from defaults
plus there was a little bug in ModOptionChanged due to it calling
updateBaseratesBasedOnPlaythrough before return, i.e. before the
CurrentValue of the slider updated, so it changed baserate value
for current playthrough to old value. Added a param to it to be
able to pass the new value instead of using the slider's
CurrentValue in that case.